+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+
+### Fixed
+- Avoid crash if tracemalloc not found, needed to run sisyphus in pypy
+
 ## [1.0.0] - 2020-02-11
 
 First release using a versioning scheme.